Rename the main package.
